Convert _project_line_string from C++ to Cython.
